+++ src/br/jus/tream/action/ActionDataEleicao.java
@@ -29,6 +29,7 @@ public class ActionDataEleicao extends ActionSupport{
 	private DataEleicao eleicao;
 	private BeanResult result;
 	private final static DataEleicaoDAO dao = DataEleicaoDAOImpl.getInstance();
+	private final static Permissao permissao = Permissao.getInstance();
 	
 	@Action(value = "listar", results = { @Result(name = "success", location = "/consultas/data-eleicao.jsp"),
 			@Result(name = "error", location = "/result.jsp")}, interceptorRefs = @InterceptorRef("authStack")
@@ -78,9 +79,9 @@ public class ActionDataEleicao extends ActionSupport{
 	public String doSetContexto() {
 		BeanResult beanResult = new BeanResult();
 		try {
-			HttpSession session = ServletActionContext.getRequest().getSession(true);
-	    	BeanLogin b = (BeanLogin)session.getAttribute("login");
-		    if (b.getAdmin()==1 && b.getZona()==0) {
+			//HttpSession session = ServletActionContext.getRequest().getSession(true);
+	    	//BeanLogin b = (BeanLogin)session.getAttribute("login");
+		    if (permissao.getAdmin()) {
 		    	beanResult.setRet(dao.ativar(this.eleicao.getId()));
 				beanResult.setMensagem(getText("eleicao.setcontexto"));
 		    } else {
@@ -100,11 +101,16 @@ public class ActionDataEleicao extends ActionSupport{
 	public String doAdicionar() {
 		BeanResult beanResult = new BeanResult();
 		try {
-			beanResult.setRet(dao.inserir(eleicao));
-			if (beanResult.getRet() == 1)
-				beanResult.setMensagem(getText("inserir.sucesso"));
-			else
-				beanResult.setMensagem(getText("inserir.error"));
+			if (permissao.getAdmin()) {
+				beanResult.setRet(dao.inserir(eleicao));
+				if (beanResult.getRet() == 1)
+					beanResult.setMensagem(getText("inserir.sucesso"));
+				else
+					beanResult.setMensagem(getText("inserir.error"));
+			}else {
+				beanResult.setRet(0);
+				beanResult.setMensagem(getText("permissao.negada"));
+			}
 		} catch (Exception e) {
 			    addActionError(getText("alterar.error") + " Error: " + e.getMessage());
 			  //result.setMensagem(getText("inserir.error") + " Error: " + e.getMessage());
@@ -119,12 +125,17 @@ public class ActionDataEleicao extends ActionSupport{
 			@Result(name = "error", location = "/pages/resultAjax.jsp") }, interceptorRefs = @InterceptorRef("authStack"))
 	public String doAtualizar() {
 		BeanResult beanResult = new BeanResult();
-		try {			
-			beanResult.setRet(dao.alterar(this.eleicao));
-			if (beanResult.getRet()==1) {
-				beanResult.setMensagem(getText("alterar.sucesso"));
+		try {
+			if (permissao.getAdmin()) {
+				beanResult.setRet(dao.alterar(this.eleicao));
+				if (beanResult.getRet()==1) {
+					beanResult.setMensagem(getText("alterar.sucesso"));
+				}else {
+					beanResult.setMensagem(getText("alterar.error")); 
+				}
 			}else {
-				beanResult.setMensagem(getText("alterar.error")); 
+				beanResult.setRet(0);
+				beanResult.setMensagem(getText("permissao.negada"));
 			}
 		} catch (Exception e) {
 			addActionError(getText("alterar.error") + " Error: " + e.getMessage());
@@ -139,8 +150,13 @@ public class ActionDataEleicao extends ActionSupport{
 	public String doRemover() {
 		BeanResult beanResult = new BeanResult();
 		try {
-			beanResult.setRet(dao.remover(this.eleicao));
-			beanResult.setMensagem(getText("remover.sucesso"));
+			if (permissao.getAdmin()) {
+				beanResult.setRet(dao.remover(this.eleicao));
+				beanResult.setMensagem(getText("remover.sucesso"));
+			}else {
+				beanResult.setRet(0);
+				beanResult.setMensagem(getText("permissao.negada"));
+			}	
 		} catch (Exception e) {
 			addActionError(getText("remover.error") + " Error: " + e.getMessage());
 			// r.setMensagem(getText("remover.error") + " Error: " + e.getMessage());
